# docs/openapi/gen/openapi.yaml
openapi: 3.0.3
info:
  title: writedown
  description: writedown
  version: 1.0.0
...
paths:
...
  /pages:
    post:
      tags:
        - pages
      summary: wikiページの作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/newPage'
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/page'
  /pages/{pageID}:
    get:
      tags:
        - pages
      summary: wikiページの取得
      parameters:
        - name: pageID
          in: path
          required: true
          schema:
            type: integer
            example: 2
      responses:
        '200':
          description: Retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/page'
    patch:
      tags:
        - pages
      summary: wikiページの更新
      parameters:
        - name: pageID
          in: path
          required: true
          schema:
            type: integer
            example: 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/patchPage'
      responses:
        '200':
          description: Updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/page'
    delete:
      tags:
        - pages
      summary: wikiページの削除
      parameters:
        - name: pageID
          in: path
          required: true
          schema:
            type: integer
            example: 2
      responses:
        '204':
          description: Deleted successfully
...
components:
  schemas:
...
    newPage:
      type: object
      required:
        - parentID
        - name
        - body
        - creator
      properties:
        parentID:
          type: integer
          example: 1
        name:
          type: string
          example: page
        body:
          type: string
          example: wikiページの内容
        creator:
          type: string
          example: owner
    page:
      type: object
      required:
        - id
        - path
        - name
        - body
        - created_at
        - updated_at
        - creator
      properties:
        id:
          type: integer
          example: 2
        path:
          type: string
          example: /path/to/page
        name:
          type: string
          example: page
        body:
          type: string
          example: wikiページの内容
        created_at:
          type: string
          format: date-time
          example: '2021-01-01T00:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2021-01-01T00:00:00Z'
        creator:
          type: string
          example: owner
    patchPage:
      type: object
      required:
        - id
        - name
        - body
        - creator
      properties:
        id:
          type: integer
          example: 2
        name:
          type: string
          example: page
        body:
          type: string
          example: wikiページの内容
        creator:
          type: string
          example: owner
...